% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myApp.R
\docType{class}
\name{CustomApplicationExample-class}
\alias{CustomApplicationExample}
\alias{CustomApplicationExample-class}
\title{Example custom KBC application in R}
\description{
Example custom KBC application in R
}
\section{Methods}{

\describe{
\item{\code{run()}}{Main application function, this should be overriden in child class.
\subsection{Return Value}{NULL}}
}}

